+++ test/nodejs/loggerTests.js
@@ -85,4 +85,39 @@ describe('logger', () => {
         assert.notStrictEqual(loggedMessages.info[0].indexOf('password:', -1));
         assert.notStrictEqual(loggedMessages.info[0].indexOf('********', -1));
     });
+
+    it('should mask reg key', () => {
+        const myRegKey = 'D3548-07483-24256-24104-0863690';
+        logger.info({ regKey: myRegKey });
+        assert.strictEqual(loggedMessages.info[0].indexOf(myRegKey), -1);
+        assert.notStrictEqual(loggedMessages.info[0].indexOf('********', -1));
+    });
+
+    it('should mask nested reg key in array', () => {
+        const myRegKey = 'D3548-07483-24256-24104-0863690';
+        logger.info({
+            license: {
+                abc: 123,
+                regKey: [myRegKey],
+                def: [3, 4, 5]
+            }
+        });
+        assert.strictEqual(loggedMessages.info[0].indexOf(myRegKey), -1);
+        assert.notStrictEqual(loggedMessages.info[0].indexOf('********', -1));
+    });
+
+    it('should mask more deeply nested reg key', () => {
+        const myRegKey = 'D3548-07483-24256-24104-0863690';
+        logger.info({
+            device: {
+                license: {
+                    foo: 'bar',
+                    regKey: myRegKey,
+                    thomas: 'jefferson'
+                }
+            }
+        });
+        assert.strictEqual(loggedMessages.info[0].indexOf(myRegKey), -1);
+        assert.notStrictEqual(loggedMessages.info[0].indexOf('********', -1));
+    });
 });